Refactors util files.
